code refactor and better status codes

+++ controllers/user.controller.js
@@ -1,3 +1,4 @@
+const { ValidationError } = require("sequelize");
 const { User } = require("../models/index");
 const bcrypt = require("bcrypt");
 const SALT_ROUNDS = 10;
@@ -29,24 +30,24 @@ exports.create = async (req, res) => {
       return res.status(422).send({
         message: "Email has already been taken.",
       });
-    } else {
-      const newUser = await User.create(user);
-      return res.send(newUser);
     }
+    const newUser = await User.create(user);
+    return res.send(newUser);
   } catch (err) {
+    console.log(err.message);
     return res.status(500).send({
-      message: err.message || "Some error occurred while creating new user.",
+      message: "Internal Server Error",
     });
   }
 };
 
 exports.findAll = async (req, res) => {
   try {
-    const users = await User.findAll();
-    res.send(users);
+    res.send(await User.findAll());
   } catch (err) {
+    console.log(err.message);
     res.status(500).send({
-      message: err.message || "Some error occurred while looking for users.",
+      message: "Internal Server Error.",
     });
   }
 };
@@ -54,16 +55,21 @@ exports.findAll = async (req, res) => {
 exports.findOne = async (req, res) => {
   try {
     const id = req.params.id;
-    const user = await User.findByPk(id);
-    if (user) return res.send(user);
-    else {
-      return res.status(404).send({
-        message: `Cannot find user with id: ${id}`,
+    if (isNaN(id) || id === " " || id === "") {
+      return res.status(400).send({
+        message: "Bad request",
       });
     }
+    const user = await User.findByPk(id);
+    if (user) return res.send(user);
+
+    return res.status(404).send({
+      message: `Cannot find user with id: ${id}`,
+    });
   } catch (err) {
+    console.log(err.message);
     return res.status(500).send({
-      message: err.message || "Some error occurred while looking for user.",
+      message: "Internal Server Error.",
     });
   }
 };
@@ -71,21 +77,47 @@ exports.findOne = async (req, res) => {
 exports.update = async (req, res) => {
   try {
     const id = req.params.id;
-    const response = await User.update(req.body, {
-      where: { id: id },
+    if (isNaN(id) || id === " " || id === "") {
+      return res.status(400).send({
+        message: "Bad request",
+      });
+    }
+    const oldUser = await User.findByPk(id);
+    if (!oldUser) {
+      return res.status(404).send({
+        message: "User not found.",
+      });
+    }
+    const newPassword = req.body.password
+      ? bcrypt.hashSync(req.body.password, SALT_ROUNDS)
+      : oldUser.password;
+    const userBody = {
+      firstName: req.body.firstName ? req.body.firstName : oldUser.firstName,
+      lastName: req.body.lastName ? req.body.lastName : oldUser.lastName,
+      email: req.body.email ? req.body.email : oldUser.email,
+      password: newPassword,
+    };
+    const response = await User.update(userBody, {
+      where: { id },
     });
-    if (response == 1) {
+    if (response) {
       return res.send({
         message: "User was updated successfully.",
       });
-    } else {
-      return res.send({
-        message: `Cannot update User with id=${id}. Maybe User was not found or req.body is empty!`,
+    }
+
+    return res.status(404).send({
+      message: `Cannot update User with id=${id}. Maybe User was not found or req.body is empty!`,
+    });
+  } catch (err) {
+    console.log(err.message);
+    if (err instanceof ValidationError) {
+      return res.status(422).send({
+        message: err.errors[0].message,
       });
     }
-  } catch (err) {
     return res.status(500).send({
-      message: "Error updating User with id=" + id,
+      message: "Internal Server Error.",
     });
   }
 };
@@ -93,21 +125,26 @@ exports.update = async (req, res) => {
 exports.destroy = async (req, res) => {
   try {
     const id = req.params.id;
+    if (isNaN(id) || id === " " || id === "") {
+      return res.status(400).send({
+        message: "Bad request",
+      });
+    }
     const response = await User.destroy({
-      where: { id: id },
+      where: { id },
     });
-    if (response == 1) {
+    if (response) {
       return res.send({
         message: "User was deleted successfully!",
       });
-    } else {
-      return res.send({
-        message: `Cannot delete User with id=${id}. Maybe User was not found!`,
-      });
     }
+    return res.status(404).send({
+      message: `User with id: ${id} not found.`,
+    });
   } catch (err) {
+    console.log(err.message);
     return res.status(500).send({
-      message: "Could not delete User with id=" + id,
+      message: "Internal Server Error.",
     });
   }
 };
